move relationship methods into DS.Model directly
